backend/requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Migrate==4.0.5
bcrypt==4.1.2
python-slugify==8.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
marshmallow==3.20.1
pytest==7.4.4
